\name{calcKinrespY}
\alias{calcKinrespY}
\title{calcKinrespY}
\description{Convert yield biomass/respiration to biomass/consumed substrate.	}
\usage{
calcKinrespY(YCO2 = 1.5)
}
\arguments{
  \item{YCO2}{
}
}



\author{Thomas Wutzler <thomas.wutzler@web.de>}



\seealso{\code{\link{twKinresp}}}

