\name{coef.kinresp}
\alias{coef.kinresp}
\title{coef kinresp}
\description{Make coef deliver attribute r0 if this is supplied with model.}
\usage{\method{coef}{kinresp}(object, ...)}
\arguments{
  \item{object}{
}
  \item{\dots}{
}
}



\author{Thomas Wutzler <thomas.wutzler@web.de>}



\seealso{\code{\link{coefKinresp.default}}
,\code{\link{twKinresp}}}

