\name{coefKinrespBeta}
\alias{coefKinrespBeta}
\title{coefKinrespBeta}
\description{Generic method behind \code{coefKinrespBeta.default}}
\usage{coefKinrespBeta(...)}
\arguments{
  \item{\dots}{
}
}



\author{Thomas Wutzler <thomas.wutzler@web.de>}




\examples{TRUE}
