\name{coefKinrespMatrix}
\alias{coefKinrespMatrix}
\title{coefKinrespMatrix}
\description{Check the microbial coefficients and translates to original microbial scale.}
\usage{coefKinrespMatrix(tmp.coef)}
\arguments{
  \item{tmp.coef}{multiple rows of microbial coefficients (see \code{\link{coefKinresp.default}})}
}



\author{Thomas Wutzler <thomas.wutzler@web.de>}



\seealso{\code{\link{coefKinresp.default}}
,\code{\link{twKinresp}}}

