\name{coefList.kinrespList}
\alias{coefList.kinrespList}
\title{coefList kinrespList}
\description{Extract the coefficients for each replicate}
\usage{\method{coefList}{kinrespList}(tmp.coef, rds.e = NULL, ...)}
\arguments{
  \item{tmp.coef}{result of \code{\link{kinrespGrowthphaseExperiment}}}
  \item{rds.e}{constrained dataset, which may omit some replicates}
  \item{\dots}{
}
}



\author{Thomas Wutzler <thomas.wutzler@web.de>}



\seealso{\code{\link{coefKinresp.default}}
,\code{\link{twKinresp}}}
\examples{FALSE}
