\name{confint.kinresp}
\alias{confint.kinresp}
\title{confint kinresp}
\description{Make confint deliver attribute r0 if this was supplied with object}
\usage{
confint.kinresp(object, ...)
}
\arguments{
  \item{object}{
}
  \item{\dots}{
}
}



\author{Thomas Wutzler <thomas.wutzler@web.de>}



\seealso{\code{\link{coefKinresp.default}}
,\code{\link{twKinresp}}}

