\name{confint.nlme}
\alias{confint.nlme}
\title{confint nlme}
\description{confidence intervals for the parameters of nlme fit, population level. }
\usage{
confint.nlme(object, ...)
}
\arguments{
  \item{object}{
}
  \item{\dots}{
}
}



\author{Thomas Wutzler <thomas.wutzler@web.de>}





