\name{confintKinresp}
\alias{confintKinresp}
\title{confintKinresp}
\description{Translate the confidence interval of estimated coefficients to original scale.}
\usage{
confintKinresp(tmp.cf)
}
\arguments{
  \item{tmp.cf}{confidence interval from fitting the microbial model \code{confint(modelfit)}}
}



\author{Thomas Wutzler <thomas.wutzler@web.de>}



\seealso{\code{\link{coefKinresp.default}}
,\code{\link{twKinresp}}}

