\name{extractCorrMatrix.gnls}
\alias{extractCorrMatrix.gnls}
\title{extractCorrMatrix gnls}
\description{Access to correlation matrix of gnls.}
\usage{
extractCorrMatrix.gnls(modelfit, ...)
}
\arguments{
  \item{modelfit}{result of gnls}
  \item{\dots}{
}
}



\author{Thomas Wutzler <thomas.wutzler@web.de>}





