\name{extractCorrMatrix.nlme}
\alias{extractCorrMatrix.nlme}
\title{extractCorrMatrix nlme}
\description{Access to correlation matrix of nlme.}
\usage{\method{extractCorrMatrix}{nlme}(modelfit, ...)}
\arguments{
  \item{modelfit}{result of nlme}
  \item{\dots}{
}
}



\author{Thomas Wutzler <thomas.wutzler@web.de>}




\examples{FALSE}
