\name{extractCorrMatrix}
\alias{extractCorrMatrix}
\title{extractCorrMatrix}
\description{Generic method behind \code{extractCorrMatrix.gnls}}
\usage{extractCorrMatrix(...)}
\arguments{
  \item{\dots}{
}
}



\author{Thomas Wutzler <thomas.wutzler@web.de>}




\examples{TRUE}
