\name{fitKinrespSuite}
\alias{fitKinrespSuite}
\title{fitKinrespSuite}
\description{Fit microbial form the entire suite.}
\usage{fitKinrespSuite(rds.e, repFits, lambda = 0.9, YCO2 = 1.5, weights = NULL, 
    start = NULL, tmp.names = c("none", "x0l", "r0l", "x0l+r0l", 
        "x0l+r0l+mumaxl"))}
\arguments{
  \item{rds.e}{
}
  \item{repFits}{
}
  \item{lambda}{
}
  \item{YCO2}{
}
  \item{weights}{
}
  \item{start}{
}
  \item{tmp.names}{
}
}



\author{Thomas Wutzler <thomas.wutzler@web.de>}



\seealso{\code{\link{twKinresp}}}

