\name{fixef.gnls}
\alias{fixef.gnls}
\title{fixef gnls}
\description{Fixed coefficients of either nlme or gnls (S3 method).}
\usage{
fixef.gnls(object, ...)
}
\arguments{
  \item{object}{model, whose fixed coefficients should be accessed}
  \item{\dots}{
}
}
\details{To support similar access to fixed coefficients of nlme and gnls models.
Because coef(nlme) will give more details than fixed.effects }


\author{Thomas Wutzler <thomas.wutzler@web.de>}





