\name{fixef.kinresp}
\alias{fixef.kinresp}
\title{fixef kinresp}
\description{Make fixed.effects deliver attribute r0 if this is supplied with model.}
\usage{\method{fixef}{kinresp}(object, ...)}
\arguments{
  \item{object}{
}
  \item{\dots}{
}
}



\author{Thomas Wutzler <thomas.wutzler@web.de>}



\seealso{\code{\link{coefKinresp.default}}
,\code{\link{twKinresp}}}

