\name{getSERId}
\alias{getSERId}
\title{getSERId}
\description{Creating a unique identifier for replicates across suits and experiments.}
\usage{getSERId(rs)}
\arguments{
  \item{rs}{the dataset with each row one replicate, with columns suite, experiment, and replicate}
}



\author{Thomas Wutzler <thomas.wutzler@web.de>}





