\name{getUnlimitedGrowthData.kinresp}
\alias{getUnlimitedGrowthData.kinresp}
\title{getUnlimitedGrowthData kinresp}
\description{Extract the dataset with records of unlimited growth phase.}
\usage{\method{getUnlimitedGrowthData}{kinresp}(kinrespRes, n = kinrespRes$n["n"], 
    ...)}
\arguments{
  \item{kinrespRes}{object of class kinresp from \code{\link{kinrespGrowthphaseReplicate}}.}
  \item{n}{the number of records in the growth phase}
  \item{\dots}{
}
}



\author{Thomas Wutzler <thomas.wutzler@web.de>}



\seealso{\code{\link{kinrespGrowthphaseExperiment}}
,\code{\link{getUnlimitedGrowthData.kinrespList}}
,\code{\link{twKinresp}}}
\examples{FALSE}
