\name{getUnlimitedGrowthData}
\alias{getUnlimitedGrowthData}
\title{getUnlimitedGrowthData}
\description{Generic method behind \code{getUnlimitedGrowthData.default}}
\usage{getUnlimitedGrowthData(...)}
\arguments{
  \item{\dots}{
}
}



\author{Thomas Wutzler <thomas.wutzler@web.de>}




\examples{TRUE}
