\name{kinrespGrowthphase}
\alias{kinrespGrowthphase}
\title{kinrespGrowthphase}
\description{Confining the unlimited growth phase all replicates in the dataset.}
\usage{
kinrespGrowthphase(rd, ...)
}
\arguments{
  \item{rd}{data.frame with columns suite, experiment, replicate, resp and time}
  \item{\dots}{further arguments to \code{\link{kinrespGrowthphaseExperiment}}}
}

\value{same format as \code{\link{kinrespGrowthphaseExperiment}}}

\author{Thomas Wutzler <thomas.wutzler@web.de>}



\seealso{\code{\link{twKinresp}}}

