\name{kinrespParDist.nlme}
\alias{kinrespParDist.nlme}
\title{kinrespParDist nlme}
\description{Expected value and confidence interval of microbial parameters}
\usage{\method{kinrespParDist}{nlme}(model, ...)}
\arguments{
  \item{model}{the result of the \code{\link{fitKinrespReplicate}} or \code{\link{fitKinrespExperiment}$model}}
  \item{\dots}{currently not used}
}



\author{Thomas Wutzler <thomas.wutzler@web.de>}



\seealso{\code{\link{kinrespParDist.gnls}}
,\code{\link{coefKinresp.default}}
,\code{\link{twKinresp}}}

