\name{plotKinrespDiagnostics.default}
\alias{plotKinrespDiagnostics.default}
\title{plotKinrespDiagnostics default}
\description{Diagnostic plots for confining unlimited growth phase.}
\usage{
plotKinrespDiagnostics.default(kinrespRes, ...)
}
\arguments{
  \item{kinrespRes}{
}
  \item{\dots}{
}
}



\author{Thomas Wutzler <thomas.wutzler@web.de>}



\seealso{,\code{\link{plotKinrespDiagnostics.kinresp}}
,\code{\link{twKinresp}}}

