 \name{respWutzler10}
 % generated in inst/cumResp/quasomLabFitsResp.R     
 \docType{data}
 \alias{respWutzler10}
 \title{Measured kinetics of Microbial Respiration.}
 \description{
   Example respiration dataset of used in Wutzler et al. 2012.
 }
 \usage{respWutzler10}
 \format{ 
    A data.frame with a row of each measurement with columns
    \itemize{
    \item{resp: respiration in gC/gCsoil/hour}
    \item{time: time of the measurement (hour)}
    \item{replicate: Identifier of the measured replicate.}
    \item{experiment: Identifier of the experiment, usually identifying one treatment}
    \item{suite: identifier of for experiment suite, if further groupings over experiments are required}
    } %itemize
 }
 \references{
    Wutzler, T.; Blagodatsky, S.; Blagodatskaya, E. & Kuzyakov, Y. (2012) Soil microbial biomass and its activity estimated by kinetic respiration analysis - Statistical guidelines. Soil Biology and Biochemistry, 45, 102-112     
 }
 \keyword{datasets}
