\name{twKinresp-internal}
\title{Internal twKinresp objects}
\alias{.coefKinrespBetaLogStart}
\alias{.coefKinrespLogStart}
\alias{.confintKinrespBeta}
\alias{.extractNlmeFullVar}
\alias{.extractRandomCovMatrix}
\alias{.fitExpModel}
\alias{.fitExpModelConstr}
\alias{.fitKinExperimentBeta}
\alias{.fitKinrespReplAll}
\alias{.kinRespStatN}
\alias{.plotRepResidualArray}
\alias{.plotRepRespArray}
\alias{.plotRepRespLogArray}
\alias{.tmp.f}
\alias{.twCalcr2}
\alias{.twVarFuncCoef}
\description{Internal twKinresp objects.}
\details{These are not to be called by the user.}
\keyword{internal}