\name{twKinresp-package}
\alias{twKinresp-package}\alias{twKinresp}
\title{Fitting kinetic models to microbial respiration data.}
\description{    Estimating microbial parameters    \itemize{    \item{ x0: initial microbial biomass }    \item{ r0: initial microbial activity }    \item{ mumax: maximum growth rate of microbial community }    }    by fitting kinetic models to microbial respiration data. }
\details{\describe{\item{Model forms}{   Microbial parameters are inferred by comparing the predictions of a microbial growth model to observations.   Observations are time series of measured increasing respiration after a soil sample has been amended with substrate for microbial growth.      The microbial model is Equation 5 in Wutzler et al. 2012. It has parameters     \itemize{    \item{ x0: initial biomass }    \item{ r0: initial activity }    \item{ mumax: maximum growth rate }    }    of the microbial community responding to the applied growth substrate.        When fitting the model in his form, the uncertainty of the model coefficients are meaningful for the inferred    microbial parameters.        The model, however, can be simplified to the following equation, here called beta-form: \cr    beta0 + beta1*exp(beta2*x)        Fitting the beta-form is usually faster and sometimes more robust.    Hence it is used to compare models to constrain respiration time series to the period of unlimited growth.  }} %describe\describe{\item{Features of the twKinresp package}{\itemize{\item{ Getting the respiration data into correct format: \code{\link{setFactorKinrespData}} }\item{ Confining the time series to the unlimited growth phase: \code{\link{kinrespGrowthphaseExperiment}} }\item{ Fitting the beta parameter form of the kinetic model to respiration data: \code{\link{fitKinrespBetaReplicate}} }\item{ Fitting the microbial paramter form of the kinetic model to respiration data: \code{\link{fitKinrespExperiment}} }\item{ Access the coefficients and confidence intervals of fitted models: \code{\link{coefKinresp.default}} }}}}  \describe{\item{Examples }{\itemize{\item{ \code{\link{respWutzler10}}:  Example respiration dataset of used in Wutzler et al. 2012 }\item{ \code{\link{fitKinrespExperiment}}: Includes a worked example}}%itemize}}  }%details \references{    Wutzler, T.; Blagodatsky, S.; Blagodatskaya, E. & Kuzyakov, Y. (2012) Soil microbial biomass and its activity estimated by kinetic respiration analysis - Statistical guidelines. Soil Biology and Biochemistry, 45, 102-112      }\author{Thomas Wutzler}

\keyword{ package }


