\name{extractCorrMatrix.gnls}
\alias{extractCorrMatrix.gnls}
\alias{extractCorrMatrix}
\title{extractCorrMatrix gnls}
\description{Access to correlation matrix of gnls.}
\usage{\method{extractCorrMatrix}{gnls}(modelfit, ...)}
\arguments{
  \item{modelfit}{result of gnls}
  \item{\dots}{
}
}



\author{Thomas Wutzler <thomas.wutzler@web.de>}





