\name{getUnlimitedGrowthData.default}
\alias{getUnlimitedGrowthData.default}
\title{getUnlimitedGrowthData default}
\description{Extract the dataset with records of unlimited growth phase.}
\usage{\method{getUnlimitedGrowthData}{default}(kinrespRes, 
    ...)}
\arguments{
  \item{kinrespRes}{
}
  \item{\dots}{
}
}



\author{Thomas Wutzler <thomas.wutzler@web.de>}



\seealso{\code{\link{getUnlimitedGrowthData.kinresp}}
\code{\link{getUnlimitedGrowthData.kinrespList}}}

