\name{getUnlimitedGrowthData}
\alias{getUnlimitedGrowthData}
\title{getUnlimitedGrowthData}
\description{Generic method behind \code{\link{getUnlimitedGrowthData.default}}}
\usage{
getUnlimitedGrowthData(...)
}
\arguments{
  \item{\dots}{
}
}



\author{Thomas Wutzler <thomas.wutzler@web.de>}





