\name{plotKinrespDiagnostics}
\alias{plotKinrespDiagnostics}
\title{plotKinrespDiagnostics}
\description{Generic method behind \code{\link{plotKinrespDiagnostics.default}}}
\usage{
plotKinrespDiagnostics(...)
}
\arguments{
  \item{\dots}{
}
}



\author{Thomas Wutzler <thomas.wutzler@web.de>}





